requirements.txt
python-telegram-bot==20.7
openai==1.58.1
python-dotenv==1.0.0
